requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
composio_core>=0.7.15
composio_openai>=0.7.15